chore: move ConfigWrapper to a helper module
